requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2
pytz>=2023.3
SQLAlchemy>=2.0.20
fastapi>=0.100.0
uvicorn>=0.23.2
pydantic>=2.3.0
